modified tests per discussion; some other requests from PR
