+++ tests/GIN_converter_spec.yml
@@ -1,47 +1,56 @@
-global_metadata:
+metadata:
   NWBFile:
     lab: My Lab
     institution: My Institution
+data_interfaces:
+  - SpikeGLXRecordingInterface
 
 experiments:
   ymaze:
-    experiment_metadata:
+    metadata:
       NWBFile:
         session_description: Subject navigating a Y-shaped maze.
     data_interfaces:
-      - SpikeGLXRecordingInterface
+      - SpikeGLXLFPInterface
 
     sessions:
-      - nwbfile_path: example_converter_spec_1.nwb
+      - nwbfile_name: example_converter_spec_1
         source_data:
           SpikeGLXRecordingInterface:
-            file_path: ephy_testing_data/spikeglx/Noise4Sam_g0/Noise4Sam_g0_imec0/Noise4Sam_g0_t0.imec0.ap.bin
-        session_metadata:
+            file_path: ephy_testing_data/Noise4Sam_g0/ Noise4Sam_g0_imec0/Noise4Sam_g0_t0.imec0.ap.bin
+          SpikeGLXLFPInterface:
+            file_path: ephy_testing_data/Noise4Sam_g0/ Noise4Sam_g0_imec0/Noise4Sam_g0_t0.imec0.lf.bin
+        conversion_options:
+          SpikeGLXRecordingInterface:
+            stub_test: True  # Required for all SpikeGLX here. File on GIN is still huge
+          BlackrockSortingExtractorInterface:
+        metadata:
           NWBFile:
-            session_start_time: "2020-11-09T21:19:09+00:00"
+            session_start_time: 2020-11-09T21:19:09+00:00
           Subject:
-            subject_id: "001"
-      - nwbfile_path: example_converter_spec_2.nwb
-        session_metadata:
+            subject_id: "1"
+      - nwbfile_path: example_converter_spec_2
+        metadata:
           NWBFile:
-            session_start_time: "2020-11-10T21:19:09+00:00"
+            session_start_time: 2020-11-10T21:19:09+00:00
           Subject:
             subject_id: "002"
         source_data:
-          example_recording:
-          example_sorting:
+          SpikeGLXRecordingInterface:
+            file_path: ephy_testing_data/Noise4Sam_g0/ Noise4Sam_g0_imec0/Noise4Sam_g0_t0.imec0.ap.bin
 
   open_explore:
-    interfaces:
-      - name: example_recording
-        data_interface: RecordingTutorialInterface
-  
     sessions:
-      - nwbfile_path: example_converter_spec_3.nwb
+      - nwbfile_name: example_converter_spec_3
+        data_interfaces:
+          - PhySortingInterface
         source_data:
-            example_recording:
-        session_metadata:
+          SpikeGLXRecordingInterface:
+            file_path: ephy_testing_data/Noise4Sam_g0/ Noise4Sam_g0_imec0/Noise4Sam_g0_t0.imec0.ap.bin
+          PhySortingInterface:
+            filename: ephy_testing_data/blackrock/FileSpec2.3001.nev
+        metadata:
           NWBFile:
-            session_start_time: "2020-11-10T21:19:09+00:00"
+            session_start_time: 2020-11-10T21:19:09+00:00
           Subject:
-            subject_id: "002"
+            subject_id: Subject Name
