+++ tests/test_tools.py
@@ -3,9 +3,7 @@ import numpy as np
 from tempfile import mkdtemp
 from pathlib import Path
 from shutil import rmtree
-from datetime import datetime
 
-from pynwb import NWBHDF5IO
 from pynwb.base import ProcessingModule
 from spikeextractors import NumpyRecordingExtractor
 
@@ -14,7 +12,6 @@ from nwb_conversion_tools.utils.conversion_tools import (
     get_module,
     make_nwbfile_from_metadata,
     estimate_recording_conversion_time,
-    convert_from_yaml,
 )
 
 
@@ -56,17 +53,3 @@ class TestConversionTools(TestCase):
         estimated_write_time, estimated_write_speed = estimate_recording_conversion_time(
             recording=recording, write_kwargs=dict(compression=None)
         )
-
-    def test_run_conversion_from_yaml(self):
-        yaml_file_path = "example_converter_spec.yml"
-        convert_from_yaml(file_path=yaml_file_path)
-        with NWBHDF5IO(path="example_converter_spec_1.nwb", mode="r") as io:
-            nwbfile = io.read()
-            assert nwbfile.session_description == "Subject navigating a Y-shaped maze."
-            assert nwbfile.lab == "My Lab"
-            assert nwbfile.institution == "My Institution"
-            assert nwbfile.session_start_time == datetime.fromisoformat("2020-11-09T21:19:09+00:00")
-            assert nwbfile.Subject.subject_id == "001"
-            assert "ElectricalSeries_raw" in nwbfile.acquisition
-            # Add more
-        # Add more files
